% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Owen.R
\name{OwenQ}
\alias{OwenQ}
\title{Owen Q-integral}
\usage{
OwenQ(nu, t, delta, a, b)
}
\arguments{
\item{nu}{finite positive number, the number of degrees of freedom}

\item{t}{finite number, positive or negative}

\item{delta}{finite number, positive or negative}

\item{a}{finite positive number, the lower bound of the integral}

\item{b}{finite positive number, the upper bound of the integral}
}
\value{
A number between \eqn{0} and \eqn{1}, the value of the integral
from \eqn{a} to \eqn{b}.
}
\description{
Evaluates the Owen Q-integral from \eqn{a} to \eqn{b} by numerical integration.
}
\examples{
# OwenQ(nu, t, delta, 0, b) = OwenQ1(nu, t, delta, b) (only for integer nu)
OwenQ(4, 3, 2, 0, 6) - OwenQ1(4, 3, 2, 6)
# failure of OwenQ for high value of nu :
nu=300; t=150; delta=160; a=0; b=100
OwenQ(nu, t, delta, a, b)
OwenQ1(nu, t, delta, b)
}
