% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Owen.R
\name{OwenQ1}
\alias{OwenQ1}
\title{First Owen Q-function}
\usage{
OwenQ1(nu, t, delta, R)
}
\arguments{
\item{nu}{integer greater than \eqn{1}, the number degrees of freedom}

\item{t}{finite number, positive or negative}

\item{delta}{finite number, positive or negative}

\item{R}{finite positive number, the upper bound of the integral}
}
\value{
A number between 0 and 1, the value of the integral from \eqn{0} to \eqn{R}.
}
\description{
Evaluates the first Owen Q-function (integral from \eqn{0} to \eqn{R})
for an integer value of the degrees of freedom.
}
\examples{
OwenQ1(4, 3, 2, 1)
# OwenQ(nu, t, delta, Inf) = pStudent(t, nu, delta)
OwenQ1(4, 3, 2, 100)
pStudent(3, 4, 2)
}
