% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OwenR.R
\name{OwenQ1_R}
\alias{OwenQ1_R}
\title{First Owen Q-function}
\usage{
OwenQ1_R(nu, t, delta, R)
}
\arguments{
\item{nu}{integer greater than \eqn{1}, the number of degrees of freedom}

\item{t}{finite number, positive or negative}

\item{delta}{finite number, positive or negative}

\item{R}{finite positive number, the upper bound of the integral}
}
\value{
A number between \eqn{0} and \eqn{1}, the value of the integral from \eqn{0} to \eqn{R}.
}
\description{
Evaluates the first Owen Q-function (integral from \eqn{0} to \eqn{R})
for an integer value of the degrees of freedom.
}
\examples{
# OwenQ1(nu, t, delta, Inf) = pStudent(t, nu, delta)
OwenQ1_R(4, 3, 2, 100)
pStudent(3, 4, 2)
}
