% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Owen.R
\name{OwenQ2}
\alias{OwenQ2}
\title{Second Owen Q-function}
\usage{
OwenQ2(nu, t, delta, R)
}
\arguments{
\item{nu}{integer greater than \eqn{1}, the number of degrees of freedom}

\item{t}{finite number, positive or negative}

\item{delta}{finite number, positive or negative}

\item{R}{finite positive number, the lower bound of the integral}
}
\value{
A number between \eqn{0} and \eqn{1}, the value of the integral
from \eqn{R} to \eqn{\infty}.
}
\description{
Evaluates the first Owen Q-function
(integral from \eqn{R} to \eqn{\infty}).
for an integer value of the degrees of freedom.
}
\examples{
OwenQ2(4, 3, 2, 1)
# OwenQ2(nu, t, delta, 0) = pStudent(t, nu, delta)
OwenQ2(4, 3, 2, 0) == pStudent(3, 4, 2)
}
