% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Owen.R
\name{pStudent}
\alias{pStudent}
\title{Student CDF with integer number of degrees of freedom}
\usage{
pStudent(q, nu, delta = 0)
}
\arguments{
\item{q}{quantile}

\item{nu}{integer greater than 1, the number of degrees of freedom}

\item{delta}{noncentrality parameter}
}
\value{
Numeric value, the CDF evaluated at \code{q}.
}
\description{
Cumulative distribution function of the noncentrel Student
distribution with an integer number of degrees of freedom.
}
\examples{
pStudent(2, 3) - pt(2, 3)
pStudent(2, 3, delta=1) - pt(2, 3, ncp=1)
}
