% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Owen.R
\name{OwenT}
\alias{OwenT}
\title{Owen T-function}
\usage{
OwenT(h, a)
}
\arguments{
\item{h}{numeric scalar}

\item{a}{numeric scalar}
}
\value{
A number between 0 and 1.
}
\description{
Evaluates the Owen T-function by numerical integration.
}
\examples{
# theoretically 0:
a <- runif(1, -1000, 1000)
OwenT(0,a) - atan(a)/(2*pi)
h <- runif(1, -3, 3)
OwenT(h,1) - pnorm(h)*(1-pnorm(h))/2
a <- 1000 # a -> Inf
OwenT(h,a) - (1-pnorm(abs(h)))/2
}
